HHH-14707 Move to Gradle's built-in way of testing Java modules
