+++ hibernate-integrationtest-java-modules/hibernate-integrationtest-java-modules.gradle
@@ -7,57 +7,22 @@
 
 description = 'Integration tests for running Hibernate ORM in the Java module path'
 
-buildscript {
-	repositories {
-		maven {
-			url "https://plugins.gradle.org/m2/"
-		}
-	}
-	dependencies {
-		classpath "org.javamodularity:moduleplugin:1.5.0"
-	}
-}
-
 apply from: rootProject.file( 'gradle/java-module.gradle' )
 
-// No first-class, built-in support for Java modules in Gradle yet,
-// so we have to use https://github.com/java9-modularity/gradle-modules-plugin
-apply plugin: "org.javamodularity.moduleplugin"
-
-// In this module, the "main" code is actually just test code that happens
-// to be built independently so as to generate a Java module.
-// So, let's override settings for compilation of the main code, just for this particular case.
-def testJavaVersions = gradle.ext.javaVersions.test
-tasks.compileJava {
-	if ( !gradle.ext.javaToolchainEnabled ) {
-		sourceCompatibility = JavaVersion.toVersion( testJavaVersions.release )
-		targetCompatibility = JavaVersion.toVersion( testJavaVersions.release )
-	}
-	else {
-		javaCompiler = javaToolchains.compilerFor {
-			languageVersion = testJavaVersions.compiler
-		}
-
-		// Remove JDK8-only options (if any) that are incompatible with options.release
-		for ( it = options.compilerArgs.listIterator(); it.hasNext(); ) {
-			if ( it.next() in ['-source', '-target'] ) {
-				it.remove()
-				it.next()
-				it.remove()
-			}
-		}
-
-		options.release = testJavaVersions.release.asInt()
-	}
+// See https://docs.gradle.org/6.7.1/userguide/java_testing.html#blackbox_integration_testing
+// See https://docs.gradle.org/6.7.1/samples/sample_java_modules_multi_project_with_integration_tests.html
+java {
+	modularity.inferModulePath = true
 }
 
 // Checkstyle fails for module-info
 checkstyleMain.exclude '**/module-info.java'
 
 dependencies {
-	compile( project( ':hibernate-core' ) )
-	compile( project( ':hibernate-envers' ) )
-	compile( libraries.jpa )
+	testCompile( project( ':hibernate-core' ) )
+	testCompile( project( ':hibernate-envers' ) )
+	testCompile( libraries.jpa )
+	testCompile( libraries.junit )
 }
 
 test {
