+++ hibernate-integrationtest-java-modules/src/test/java/org/hibernate/orm/integrationtest/java/module/test/JavaModulePathIT.java
@@ -11,7 +11,7 @@ import java.util.Arrays;
 
 import org.hibernate.Session;
 import org.hibernate.envers.boot.internal.EnversIntegrator;
-import org.hibernate.orm.integrationtest.java.module.service.AuthorService;
+import org.hibernate.orm.integrationtest.java.module.test.service.AuthorService;
 
 import org.junit.Assert;
 import org.junit.Test;
