+++ hibernate-integrationtest-java-modules/src/test/java/org/hibernate/orm/integrationtest/java/module/test/entity/Author.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.orm.integrationtest.java.module.entity;
+package org.hibernate.orm.integrationtest.java.module.test.entity;
 
 import javax.persistence.Basic;
 import javax.persistence.Entity;
